refactor(docx): unify table row-height rule into one path (§17.4.80/.85)
The pt paginator path (computeTableRowHeights) and the px paint path
(computeTableLayout / calculateRowHeight) each re-implemented the same
ST_HeightRule skeleton (§17.4.80/§17.18.37: exact/atLeast/auto floor +
MIN_ROW_HEIGHT_PT), gridSpan column slicing, vMerge restart-cell
exclusion, and the §17.4.85 vMerge span post-pass. #523 had to patch the
identical regression in both places; future single-side edits would drift
measure and paint apart.

Extract that rule+span skeleton into a new table-geometry.ts
(resolveTableRowHeights / resolveSingleRowHeight, MIN_ROW_HEIGHT_PT,
findMergeEndRow). Each caller now supplies only its unit-specific cell
content measurer via a callback:
  - pt path: scale=1, estimateParagraphHeight cursor-walk (unchanged)
  - px path: scale=device, measureCellElementHeight (unchanged)

The two measurers are intentionally NOT unified — they are not
equivalent (e.g. an empty East-Asian paragraph mark on a docGrid rounds
to a different cell count between estimateParagraphHeight and
measureParaHeight; also first-line/left-right indent and float-band
displacement). Picking one would change rendering and needs separate
Word/spec adjudication. This change is rule-skeleton de-duplication only:
the per-cell measurement, and therefore all output, is byte-identical.

Behavior unchanged: table-row-height.test.ts /
measure-column-geometry.test.ts and the full docx suite (195 tests) pass
with no expected-value edits; root pnpm typecheck clean.
